+++ examples/using-gatsby-image/src/pages/index.js
@@ -1,5 +1,5 @@
 import React from "react"
-import { Link } from "gatsby"
+import { Link, graphql } from "gatsby"
 import Img from "gatsby-image"
 import Navigation from "../components/navigation"
 import Layout from "../layouts"
@@ -120,7 +120,7 @@ class IndexComponent extends React.Component {
 export default IndexComponent
 
 export const query = graphql`
-  query FrontPageQuery {
+  query {
     file(relativePath: { regex: "/ng-55646/" }) {
       childImageSharp {
         fluid(maxWidth: 1500, rotate: 180) {