+++ examples/using-javascript-transforms/src/mainPages/contact.js
@@ -1,9 +1,7 @@
 import React from "react"
-import Helmet from "react-helmet"
-import SiteLinks from "../components/SiteLinks"
 import InsetPageLayout from "../components/Layouts/insetPage"
 
-exports.frontmatter = {
+export const frontmatter = {
   layoutType: `page`,
   path: `/contact/`,
 }
@@ -23,7 +21,7 @@ class ContactMe extends React.Component {
 export default ContactMe
 
 export const pageQuery = graphql`
-  query contactUs {
+  query {
     site {
       ...site_sitemetadata
     }