+++ packages/gatsby-plugin-less/src/gatsby-node.js
@@ -30,11 +30,11 @@ exports.onCreateWebpackConfig = (
   const lessRuleModules = {
     test: /\.module\.less$/,
     use: [
-      loaders.miniCssExtract(),
+      !isSSR && loaders.miniCssExtract(),
       loaders.css({ modules: true, importLoaders: 2 }),
       loaders.postcss({ plugins: postCssPlugins }),
       lessLoader,
-    ],
+    ].filter(Boolean),
   }
 
   let configRules = []