+++ packages/gatsby-plugin-netlify-cms/src/gatsby-node.js
@@ -1,74 +1,143 @@
-const HtmlWebpackPlugin = require(`html-webpack-plugin`)
-const ExtractTextPlugin = require(`extract-text-webpack-plugin`)
-
-const extractCmsCss = new ExtractTextPlugin(`cms.css`)
-
-function plugins(stage) {
-  const commonPlugins = [
-    // Output /admin/index.html
-    new HtmlWebpackPlugin({
-      title: `Content Manager`,
-      filename: `admin/index.html`,
-      chunks: [`cms`],
-    }),
-  ]
-
-  switch (stage) {
-    case `develop`:
-      return commonPlugins
-    case `build-javascript`:
-      return [...commonPlugins, extractCmsCss]
-    default:
-      return []
-  }
-}
+import path from "path"
+import { get, mapValues, isPlainObject, trim, pickBy } from "lodash"
+import webpack from "webpack"
+import HtmlWebpackPlugin from "html-webpack-plugin"
+import MiniCssExtractPlugin from "mini-css-extract-plugin"
+import UglifyJsPlugin from "uglifyjs-webpack-plugin"
+import FriendlyErrorsPlugin from "friendly-errors-webpack-plugin"
 
 /**
- * Exclude Netlify CMS styles from Gatsby CSS bundle. This relies on Gatsby
- * using webpack-configurator for webpack config extension, and also on the
- * target loader key being named "css" in Gatsby's webpack config.
+ * Deep mapping function for plain objects and arrays. Allows any value,
+ * including an object or array, to be transformed.
  */
-function excludeFromLoader(key, config) {
-  config.loader(key, ({ exclude, ...configRest }) => {
-    const regex = /\/node_modules\/netlify-cms\//
-    if (!exclude) {
-      return { ...configRest, exclude: regex }
-    }
-    if (Array.isArray(exclude)) {
-      return { ...configRest, exclude: [...exclude, regex] }
-    }
-    return { ...configRest, exclude: [exclude, regex] }
-  })
-}
-
-function module(config, stage) {
-  switch (stage) {
-    case `build-css`:
-      excludeFromLoader(`css`, config)
-      return config
-    case `build-javascript`:
-      excludeFromLoader(`css`, config)
-
-      // Exclusively extract Netlify CMS styles to /cms.css (filename configured
-      // above with plugin instantiation).
-      config.loader(`cms-css`, {
-        test: /\.css$/,
-        include: [/\/node_modules\/netlify-cms\//],
-        loader: extractCmsCss.extract([`css`]),
-      })
-      return config
-    default:
-      return config
+function deepMap(obj, fn) {
+  /**
+   * If the transform function transforms the value, regardless of type,
+   * return the transformed value.
+   */
+  const mapped = fn(obj)
+  if (mapped !== obj) {
+    return mapped
   }
+
+  /**
+   * Recursively deep map arrays and plain objects, otherwise return the value.
+   */
+  if (Array.isArray(obj)) {
+    return obj.map(value => deepMap(value, fn))
+  }
+  if (isPlainObject(obj)) {
+    return mapValues(obj, value => deepMap(value, fn))
+  }
+  return obj
 }
 
-exports.onCreateWebpackConfig = ({ actions, stage }, { modulePath }) => {
-  const config = actions.setWebpackConfig({
-    entry: {
-      cms: [`${__dirname}/cms.js`, modulePath].filter(p => p),
-    },
-    plugins: plugins(stage),
-  })
+exports.onCreateWebpackConfig = (
+  { store, stage, getConfig, plugins },
+  {
+    modulePath,
+    stylesPath,
+    publicPath = `admin`,
+    enableIdentityWidget = true,
+    htmlTitle = `Content Manager`,
+  }
+) => {
+  if ([`develop`, `build-javascript`].includes(stage)) {
+    const gatsbyConfig = getConfig()
+    const { program } = store.getState()
+    const publicPathClean = trim(publicPath, `/`)
+    const config = {
+      ...gatsbyConfig,
+      mode: `none`,
+      /**
+       * Two entries, one for the core CMS styles, and a `styles` entry, the
+       * output of which will be applied to the preview pane iframe only. Use
+       * `pickBy` to filter out empty entries.
+       */
+      entry: pickBy({
+        cms: [
+          `${__dirname}/cms.js`,
+          modulePath,
+          enableIdentityWidget && `${__dirname}/cms-identity.js`,
+        ].filter(p => p),
+        styles: stylesPath,
+      }),
+      output: {
+        path: path.join(program.directory, `public`, publicPathClean),
+      },
+      module: {
+        /**
+         * Manually swap `style-loader` for `MiniCssExtractPlugin.loader`.
+         * `style-loader` is only used in development, and doesn't allow us to
+         * pass the `styles` entry css path to Netlify CMS.
+         */
+        rules: deepMap(gatsbyConfig.module.rules, value => {
+          if (
+            typeof get(value, `loader`) === `string` &&
+            value.loader.includes(`style-loader`)
+          ) {
+            return { ...value, loader: MiniCssExtractPlugin.loader }
+          }
+          return value
+        }),
+      },
+      plugins: [
+        /**
+         * Remove plugins that either attempt to process the core Netlify CMS
+         * application, or that we want to replace with our own instance.
+         */
+        ...gatsbyConfig.plugins.filter(
+          plugin =>
+            ![UglifyJsPlugin, MiniCssExtractPlugin, FriendlyErrorsPlugin].find(
+              Plugin => plugin instanceof Plugin
+            )
+        ),
 
-  module(config, stage)
+        /**
+         * Provide a custom message for Netlify CMS compilation success.
+         */
+        stage === `develop` &&
+          new FriendlyErrorsPlugin({
+            clearConsole: false,
+            compilationSuccessInfo: {
+              messages: [
+                `Netlify CMS is running at ${
+                  program.ssl ? `https` : `http`
+                }://${program.host}:${program.port}/${publicPathClean}/`,
+              ],
+            },
+          }),
+
+        /**
+         * Use a simple filename with no hash so we can access from source by
+         * path.
+         */
+        new MiniCssExtractPlugin({
+          filename: `[name].css`,
+        }),
+
+        /**
+         * Auto generate CMS index.html page.
+         */
+        new HtmlWebpackPlugin({
+          title: htmlTitle,
+          chunks: [`cms`],
+        }),
+
+        /**
+         * Set flag if custom styles path is added to plugin options.
+         */
+        plugins.define({
+          NETLIFY_CMS_PREVIEW_STYLES_SET: !!stylesPath,
+        }),
+      ].filter(p => p),
+
+      /**
+       * Remove common chunks style optimizations from Gatsby's default config,
+       * they cause issues for our pre-bundled code.
+       */
+      optimization: {},
+    }
+    webpack(config).run()
+  }
 }