+++ packages/gatsby-plugin-offline/src/gatsby-node.js
@@ -3,6 +3,7 @@ const precache = require(`sw-precache`)
 const path = require(`path`)
 const slash = require(`slash`)
 const _ = require(`lodash`)
+const cheerio = require(`cheerio`)
 
 exports.createPages = ({ actions }) => {
   if (process.env.NODE_ENV === `production`) {
@@ -44,12 +45,36 @@ exports.onPostBuild = (args, pluginOptions) => {
     rootDir
   )
 
+  // load index.html to pull scripts/links necessary for proper offline reload
+  const html = fs.readFileSync(
+    path.resolve(`${process.cwd()}/${rootDir}/index.html`)
+  )
+
+  // party like it's 2006
+  const $ = cheerio.load(html)
+
+  // holds any paths for scripts and links
+  const criticalFilePaths = []
+
+  $(`script`)
+    .filter((_, elem) => $(elem).attr(`src`) !== undefined)
+    .each((_, elem) => {
+      criticalFilePaths.push(`${rootDir}${$(elem).attr(`src`)}`)
+    })
+
+  $(`link`)
+    .filter((_, elem) => $(elem).attr(`as`) !== `script`)
+    .each((_, elem) => {
+      criticalFilePaths.push(`${rootDir}${$(elem).attr(`href`)}`)
+    })
+
   const options = {
     staticFileGlobs: files.concat([
       `${rootDir}/index.html`,
       `${rootDir}/manifest.json`,
       `${rootDir}/manifest.webmanifest`,
       `${rootDir}/offline-plugin-app-shell-fallback/index.html`,
+      ...criticalFilePaths,
     ]),
     stripPrefix: rootDir,
     // If `pathPrefix` is configured by user, we should replace