+++ packages/gatsby/src/cache-dir/public-page-renderer.js
@@ -1,7 +1,0 @@
-const preferDefault = m => (m && m.default) || m
-
-if (process.env.NODE_ENV === `production`) {
-  module.exports = preferDefault(require(`./public-page-renderer-prod`))
-} else {
-  module.exports = preferDefault(require(`./public-page-renderer-dev`))
-}