+++ www/gatsby-node.js
@@ -79,7 +79,7 @@ exports.createPages = ({ graphql, actions }) => {
           query {
             allMarkdownRemark(
               sort: { order: DESC, fields: [frontmatter___date] }
-              limit: 1000
+              limit: 10000
               filter: { fileAbsolutePath: { ne: null } }
             ) {
               edges {
@@ -107,7 +107,7 @@ exports.createPages = ({ graphql, actions }) => {
                 }
               }
             }
-            allSitesYaml(limit: 40, filter: { main_url: { ne: null } }) {
+            allSitesYaml(filter: { main_url: { ne: null } }) {
               edges {
                 node {
                   fields {
@@ -144,11 +144,13 @@ exports.createPages = ({ graphql, actions }) => {
           edge => {
             const slug = _.get(edge, `node.fields.slug`)
             const draft = _.get(edge, `node.frontmatter.draft`)
-            if (!slug) return
+            if (!slug) return undefined
 
             if (_.includes(slug, `/blog/`) && !draft) {
               return edge
             }
+
+            return undefined
           }
         )