+++ www/src/templates/tags.js
@@ -58,7 +58,7 @@ Tags.propTypes = {
 export default Tags
 
 export const pageQuery = graphql`
-  query TagPage($tag: String) {
+  query($tag: String) {
     allMarkdownRemark(
       limit: 2000
       sort: { fields: [frontmatter___date], order: DESC }